have gbrowse_clean.pl remove outdated uploads from user database

+++ bin/gbrowse_clean.pl
@@ -6,7 +6,7 @@ use Getopt::Long;
 use Pod::Usage;
 
 use File::Find ();
-use File::Basename 'basename';
+use File::Basename 'basename','dirname';
 use Bio::Graphics::Browser2;
 use Bio::Graphics::Browser2::Util 'shellwords';
 use POSIX 'ENOTEMPTY';
@@ -41,6 +41,7 @@ my $cache_dir     = $globals->cache_dir;
 my $locks_dir     = $globals->session_locks;
 my $images_dir    = $globals->tmpimage_dir;
 my $user_dir      = $globals->user_dir();
+my $uploads_db    = $globals->user_account_db;
 
 my $cache_secs   = $globals->time2sec($globals->cache_time);
 my $uploads_secs = $globals->time2sec($globals->upload_time);
@@ -82,6 +83,9 @@ my $wanted = sub {
 	$rdev,$size,$atime,$mtime,$ctime) = stat($_); 
 
     next if $name =~ m!$tmpdir/[^/]+$!; # don't remove toplevel!
+    next if $name eq $tmpdir;
+
+    my $is_userdata = $name =~ m/^$user_dir/;
 
     if (-d _ ) { # attempt to remove directories - will have no effect unless empty
 	if (rmdir($name)) {
@@ -94,10 +98,10 @@ my $wanted = sub {
     }
 
     my $secs = $name =~ m/^($cache_dir|$locks_dir|$images_dir)/  ? $cache_secs
-	      :$name =~ m/^$user_dir/                            ? $uploads_secs
+	      :$is_userdata                                      ? $uploads_secs
 	      :0;
     return unless $secs;
-    my $time = $name =~ m/^$user_dir/ ? -A _ : -M _;
+    my $time = $is_userdata ? -A _ : -M _;
 
     my $days = $secs/SECS_PER_DAY;
 
@@ -113,12 +117,37 @@ my $wanted = sub {
 
 # Traverse desired filesystems
 logit("Deleting cache files and directories...");
-File::Find::finddepth( {wanted=>$wanted},
-		       $tmpdir);
-logit("$directories directories and $files files deleted.\n");
+File::Find::finddepth( {wanted=>$wanted},  $tmpdir);
+logit("Deleting unused user uploads olderthan $uploads_secs seconds (see \"expire uploads\" option in GBrowse.conf)...");
+File::Find::finddepth( {wanted=>$wanted},  $user_dir);
+logit("Deleted $directories directories and $files files.\n");
+if ($uploads_db) {
+    logit("Cleaning uploads db...");
+    clean_uploads();
+}
 
 logit("*** ",scalar localtime,"$0 done ***\n\n");
 exit 0;
+
+sub clean_uploads {
+    eval {require DBI; 1} or return;
+
+    my $db = DBI->connect($globals->user_account_db) or return;
+    my $query = $db->prepare('select sessionid,data_source,path,$trackid from uploads,session where uploads.userid=session.userid') or return;
+    $query->execute or return;
+
+    my %flag_for_deletion;
+    while (my ($sessionid,$dsn,$path,$trackid) = $query->fetchrow_array) {
+	my $full_path = "$user_dir/$dsn/$path";
+	next if -e $full_path;
+	$flag_for_deletion{$trackid}++;
+    }
+
+    my @delete = keys %flag_for_deletion or return;
+    my $to_remove = join ',',map {"'$_'"}@delete;
+    logit("Deleting ".scalar @delete." dangling uploads");
+    $db->do("delete from uploads where trackid in ($to_remove)") or warn $db->errstr;
+}
 
 sub verbose {
     my @mess = @_;
